refactor: moved admin tests from non-admin test class
