+++ src/test/java/com/dsrm/dsrmbackend/RoomTypeControllerTests.java
@@ -11,7 +11,6 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.transaction.annotation.Transactional;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -59,22 +58,5 @@ public class RoomTypeControllerTests extends AbstractIntegrationTest {
                         .andExpect(jsonPath("$.content[0].name").value("Pokoj mieszkalny"))
                         .andExpect(jsonPath("$.content[1].name").value("Sala telewizyjna"));
     }
-
-    @Transactional
-    @Test
-    public void deleteExistentRoomType() throws Exception {
-        this.mockMvc.perform(delete("/admin/room-types/2")
-                .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNoContent())
-                .andExpect(jsonPath("$.name").value("Sala telewizyjna"));
-    }
-
-    @Test
-    public void tryToDeleteExistentRoomType() throws Exception {
-        this.mockMvc.perform(delete("/admin/room-types/2222")
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNotFound());
-    }
-
 
 }
